Move `discover` and `discoverSsr` to index.web.js and index.node.js
And extract DiscoverDocumentHead to a separate file.
